added validation dependencies and autowired services

+++ src/main/java/com/example/restgreeting/models/Greeting.java
@@ -3,13 +3,13 @@ package com.example.restgreeting.models;
 import static com.example.restgreeting.constants.StringConstants.REQUIRED_FIELD;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.validation.constraints.NotBlank;
 import java.util.Objects;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import javax.validation.constraints.NotBlank;
 
 @Entity
 @Table(name = "greetings")
@@ -17,7 +17,7 @@ public class Greeting {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private long id;
+  private Long id;
   @NotBlank(message = "greeting" + REQUIRED_FIELD)
   private String greeting;
 
@@ -32,7 +32,7 @@ public class Greeting {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(Long id) {
     this.id = id;
   }
